[CARBONDATA-1405]Added support for gathering coverage information when tests are run
Using jacoco for coverage
Using coveralls for CI integration

This closes #1288
